Create views-azure-en_us-lang.php-notes
Some of this will seem as strange to you as what was there was to me. I would do better in some cases knowing the context of the language files, and in other cases if I knew more about programming. These for sure will have to be supervised by a more knowledgeable person. Some changes are personal preferences. Others are to be grammatically precise. Other items are not changed at all to comply with shorthand programming lingo. I hope these files help. I am available of course for any questions.